+++ extensions/chocolatey-core.extension/extensions/Get-PackageCacheLocation.ps1
@@ -24,7 +24,10 @@ function Get-PackageCacheLocation {
         # Name of the package, by default $Env:ChocolateyPackageName
         [string] $Name    = $Env:ChocolateyPackageName,
         # Version of the package, by default $Env:ChocolateyPackageVersion
-        [string] $Version = $Env:ChocolateyPackageVersion
+        [string] $Version = $Env:ChocolateyPackageVersion,
+        # Allows splatting with arguments that do not apply and future expansion. Do not use directly.
+        [parameter(ValueFromRemainingArguments = $true)]
+        [Object[]] $IgnoredArguments
     )
 
     if (!$Name) { Write-Warning 'Environment variable $Env:ChocolateyPackageName is not set' }