+++ extensions/chocolatey-core.extension/extensions/Get-WebContent.ps1
@@ -43,7 +43,11 @@ function Get-WebContent {
         [string]$Url,
 
         # Additional options for http request.For now only Headers property supported.
-        [hashtable]$Options
+        [hashtable]$Options,
+
+        # Allows splatting with arguments that do not apply and future expansion. Do not use directly.
+        [parameter(ValueFromRemainingArguments = $true)]
+        [Object[]] $IgnoredArguments
     )
 
     $filePath =  get_temp_filepath